Clean up and store last_kicked in MongoDB

+++ src/main/java/xyz/nikitacartes/easyauth/storage/AuthConfig.java
@@ -289,7 +289,7 @@ public class AuthConfig {
          * How long it takes (seconds) after a player gets kicked
          * for too many logins for the player to be allowed back in.
          */
-        public int resetLoginAttemptsTime = 120;
+        public long resetLoginAttemptsTime = 120;
     }
 
     public MainConfig main = new MainConfig();
